+++ devcovenant/policy_scripts/line_length_limit.py
@@ -8,6 +8,7 @@ from typing import List
 
 from devcovenant.base import CheckContext, PolicyCheck, Violation
 
+
 class LineLengthLimitCheck(PolicyCheck):
     """
     Check that lines are under 79 characters.