+++ devcovenant/registry.py
@@ -11,6 +11,7 @@ from typing import Dict, List, Optional
 
 from .parser import PolicyDefinition
 
+
 @dataclass
 class PolicySyncIssue:
     """
@@ -34,6 +35,7 @@ class PolicySyncIssue:
     script_exists: bool
     issue_type: str
     current_hash: Optional[str] = None
+
 
 class PolicyRegistry:
     """